Revert "x86/retpoline: Remove .text..__x86.return_thunk section"
This reverts commit e92626af3234708fe30f53b269d210d202b95206.

David Howells reported his box freezing without being able to see
a panic. However, it managed to issue a warning beforehand:

  missing return thunk: __x86_indirect_thunk_r15+0xa/0x5f-0x0: eb 74 66 66 2e
  WARNING: CPU: 0 PID: 0 at arch/x86/kernel/alternative.c:755 apply_returns+0xca/0x247
  Modules linked in:
  CPU: 0 PID: 0 Comm: swapper/0 Not tainted 6.6.0-rc5-next-20231013-build3+ #3044
  Hardware name: ASUS All Series/H97-PLUS, BIOS 2306 10/09/2014
  RIP: 0010:apply_returns+0xca/0x247

this happened with linux-next and with gcc 13. Looking at the compiler
output and particularly paying attention to the two JMP instructions:

  <__x86_indirect_thunk_r14>:
         e8 01 00 00 00          call   ffffffff81d71206 <__x86_indirect_thunk_r14+0x6>
         cc                      int3
         4c 89 34 24             mov    %r14,(%rsp)
         e9 91 00 00 00          jmp    ffffffff81d712a0 <__x86_return_thunk>
         66 66 2e 0f 1f 84 00    data16 cs nopw 0x0(%rax,%rax,1)
         00 00 00 00
         66 0f 1f 44 00 00       nopw   0x0(%rax,%rax,1)

  <__x86_indirect_thunk_r15>:
         e8 01 00 00 00          call   ffffffff81d71226 <__x86_indirect_thunk_r15+0x6>
         cc                      int3
         4c 89 3c 24             mov    %r15,(%rsp)
         eb 74                   jmp    ffffffff81d712a0 <__x86_return_thunk>

the second JMP is a short JMP one. This is likely some new gcc
optimization to size the JMP offsets and generate a small one if it
fits.

However, the apply_returns() logic does not expect a short JMP:

  if (op == JMP32_INSN_OPCODE)
        dest = addr + insn.length + insn.immediate.value;

and that JMP32_INSN_OPCODE is 0xe9.

Now, if __x86_return_thunk is in another section, the compiler cannot do
those shortcuts and will have to generate a JMP with a s32 offset.

As a matter of fact, the removal of the section  broke another case, see

  https://lore.kernel.org/r/20231010171020.462211-2-david.kaplan@amd.com

so revert for now until all the possible code generation issues have
been assessed, addressed and verified properly.

Reported-by: David Howells <dhowells@redhat.com>
Tested-by: David Howells <dhowells@redhat.com>
Signed-off-by: Borislav Petkov (AMD) <bp@alien8.de>
Cc: Josh Poimboeuf <jpoimboe@kernel.org>
Link: https://lore.kernel.org/r/20231018175531.GEZTAcE2p92U1AuVp1@fat_crate.local