dt-bindings: clock: ti: Convert to yaml
This binding doesn't define a new clock binding type,
it is used to group the existing clock nodes under the hardware hierarchy.

As this is not a provider clock, remove #clock-cells and
clock-output-names properties.
Though few clockdomain nodes in the dts use these properties,
we are not fixing dts here.
Clean up the example to meet the current standards.

Add the creator of the original binding as a maintainer.